Refactoring
Refactor.js is now the primary JS file for the app.
It contains the #sortable IIFE that activates the
drag-and-drop functionality, as well as bindEvents().

I extracted CreateCard and addCardToDOM from addNewCard.

I made a handleDoubleClick function that fires multiple
single-responsibility functions (extracted from other
functions), such as finding the id, querying server,
filling the modal, and launching it.

I renamed findListId to createMovement because that is
what the function was actually doing.

I revised the updateCard function to explicitly reference
the database, as well as including three callback functions
for updating different parts of the DOM in the done callback.
I did this so the whole board would not need to be rendered
again every time a card was updated.

Because the code has not been tested fully yet, there are
many debuggers and console log statements in the code.